apps/api/requirements.txt
# Simon AI Agent Studio - API Dependencies
# Python 3.11+
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
# HTTP Client
httpx==0.26.0
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
# Redis
redis==5.0.1
# Celery (UI Runner integration)
celery==5.3.6
# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7
# Observability
psutil==5.9.8
# Utilities
python-dateutil==2.8.2

# Computer Use - Playwright
playwright==1.40.0
pillow==10.1.0

prometheus-client==0.19.0
prometheus-fastapi-instrumentator==7.0.0
